backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
PyMuPDF
pytesseract
Pillow
transformers
torch
openai
deep-translator
